pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "woodchipper"
version = "0.1.0"
description = "A Python library and CLI tool"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Your Name", email = "you@example.com" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "python-magic>=0.4.27",
    "pypdf>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
woodchipper = "woodchipper.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/woodchipper"
Repository = "https://github.com/yourusername/woodchipper"

[tool.hatch.build.targets.wheel]
packages = ["src/woodchipper"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
